Allow edits and move files to grid folder
